Created the Code
Completed the Assignment: Program requests the file name from the user, it uses the correct file name by default if the user leaves the input blank (<1 character), opens the file and reads through it. It looks for 'From ' lines (while skipping lines that begin with 'From:'), takes the second word of those lines, creates a Python dictionary (Key-Value / 'Sender's mail address'-'Count of the number of times they appear in the file'). It reads through the dictionary and uses a loop to count the number of times each key (email) appears, then prints the maximum (which email appears the most and its count).